fixed several issues
