+++ python-client/giskard/models/huggingface/__init__.py
@@ -1,19 +1,17 @@
 import mlflow
 from typing import Union
 import logging
-import numpy as np
 
 from giskard.core.core import SupportedModelTypes
 from giskard.core.model import MLFlowBasedModel
 
-import transformers
 from transformers import PreTrainedModel
 
 logger = logging.getLogger(__name__)
 
 
 class HuggingFaceModel(MLFlowBasedModel):
-    #TODO: add this always should_save_model_class = True
+    # TODO: add this always should_save_model_class = True
     def __init__(self,
                  clf,
                  model_type: Union[SupportedModelTypes, str],
@@ -40,7 +38,7 @@ class HuggingFaceModel(MLFlowBasedModel):
     def save_with_mlflow(self, local_path, mlflow_meta: mlflow.models.Model):
         self.clf.save_pretrained(local_path)
 
-    #TODO: abstract clf_predict (extreme plan B)
+    # TODO: abstract clf_predict (extreme plan B)
     def clf_predict(self, data):
         predictions = self.clf.predict(data)
         return predictions
