+++ python-client/pyproject.toml
@@ -104,10 +104,6 @@ tensorflow = "^2.11.0"
 tensorflow-hub = "^0.12.0"
 tensorflow-text = "^2.11.0"
 transformers = "^4.26.1"
-<<<<<<< HEAD
-nltk = "^3.8.1"
-=======
->>>>>>> feature/new-upload-api-tensorflow
 
 [tool.black]
 # https://github.com/psf/black
